refactor: reorganize agent/ and tests/agent/ into core and domain subdirectories
agent/core/   ← graph, nodes, state, model, telemetry, temporal
agent/domain/ ← triage, planner, tracker, care, weather, incidents,
                interactions, activity_log

tests/agent/core/   ← test_graph, test_nodes, test_node_edge_cases, test_telemetry
tests/agent/domain/ ← test_domain_logic

All import paths updated across agent/, main.py, tests/, and
tests/support/patching.py via find+sed. No backwards compatibility shims.
286 tests pass.
