+++ agent/__init__.py
@@ -1,1 +1,1 @@
-from agent.graph import agent
+from agent.core.graph import agent
