+++ agent/domain/interactions.py
