+++ agent/tools/garden/plants.py
@@ -3,7 +3,7 @@ Agent tools for managing plants.
 Tools must return strings — the LLM reads tool output as text.
 """
 from langchain.tools import tool
-from agent.activity_log import (
+from agent.domain.activity_log import (
     DEFAULT_ACTOR_LABEL,
     DEFAULT_ACTOR_TYPE,
     record_activity_event,
