+++ agent/tools/operations/triage.py
@@ -4,8 +4,8 @@ from typing import Optional
 
 from langchain.tools import tool
 
-from agent.temporal import DEFAULT_TIMEZONE
-from agent.triage import build_triage_snapshot, format_triage_snapshot
+from agent.core.temporal import DEFAULT_TIMEZONE
+from agent.domain.triage import build_triage_snapshot, format_triage_snapshot
 from db.database import SessionLocal
 from db.models import TriageSnapshot
