+++ tests/db/test_planner.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from agent.planner import (
+from agent.domain.planner import (
     build_execution_spec_payload,
     build_plan_input,
     check_plan_feasibility,
